backend/requirements.txt
fastapi
uvicorn
llama-index
llama-index-llms-nvidia
llama-index-embeddings-nvidia
arxiv
networkx
python-dotenv
pypdf
